chore: Move all server config routes in the same file
